+++ tests/testthat/test-simtrait.R
@@ -326,3 +326,36 @@ test_that( "pval_srmsd works", {
     expect_true( all( data$pvals_unif <= 1 ) )
     expect_true( all( data$pvals_null <= 1 ) )
 })
+
+test_that( "pval_aucpr works", {
+    # random data to test on
+    n <- 10
+    n_null <- 5 # must be strictly smaller than n for things to work
+    # all p-values are uniform
+    pvals <- runif( n )
+    # pick a few random ones to be causal (to be removed inside)
+    causal_loci <- sample.int( n, n - n_null )
+    # pick one of each class to be NA
+    pvals[ causal_loci ][1] <- NA
+    pvals[ -causal_loci ][1] <- NA
+    
+    # trigger failures on purpose
+    # (missing arguments)
+    expect_error( pval_aucpr( ) )
+    expect_error( pval_aucpr( pvals ) )
+    expect_error( pval_aucpr( causal_loci = causal_loci ) )
+    # p-values out of range cause errors
+    expect_error( pval_aucpr( c(pvals, -1), causal_loci ) )
+    expect_error( pval_aucpr( c(pvals, 10), causal_loci ) )
+    # empty causal_loci trigger a specific error
+    expect_error( pval_aucpr( pvals, c() ) )
+    # and all loci being causal (no nulls) also triggers errors
+    expect_error( pval_aucpr( pvals, 1:length(pvals) ) )
+
+    # now a successful case
+    expect_silent( auc <- pval_aucpr( pvals, causal_loci ) )
+    expect_equal( length(auc), 1 )
+    expect_true( !is.na(auc) )
+    expect_true( auc >= 0 )
+    expect_true( auc <= 1 )
+})